feat: i18n
Well, the commit started with fixing some problem with resource bundles
when installing the plugin on my own Intellij-IDEA.
Not sure what exactly the problem was, but after reading the
documentation and doing everything "by the book", it works now.
Along the way of doing things "by the book", I added also i18n.